// x-pack/solutions/observability/plugins/apm/public/embeddable/types.d.ts
import type { CoreStart, CoreSetup } from '@kbn/core/public';
import type { ObservabilityRuleTypeRegistry } from '@kbn/observability-plugin/public';
import type { ApmPluginStartDeps, ApmPluginSetupDeps } from '../plugin';
import type { ITelemetryClient } from '../services/telemetry';
import type { ConfigSchema } from '..';
import type { KibanaEnvContext } from '../context/kibana_environment_context/kibana_environment_context';
export interface EmbeddableDeps {
    coreStart: CoreStart;
    pluginsStart: ApmPluginStartDeps;
    coreSetup: CoreSetup;
    pluginsSetup: ApmPluginSetupDeps;
    config: ConfigSchema;
    kibanaEnvironment: KibanaEnvContext;
    observabilityRuleTypeRegistry: ObservabilityRuleTypeRegistry;
    telemetry: ITelemetryClient;
}
export interface APMEmbeddableProps {
    transactionName?: string;
    rangeFrom?: string;
    rangeTo?: string;
    kuery?: string;
}
export type APMEmbeddableInput = APMEmbeddableProps;